src/funcmcp/requirements.txt
# Uncomment to enable Azure Monitor OpenTelemetry
# Ref: aka.ms/functions-azure-monitor-python
# azure-monitor-opentelemetry

azure-functions
python-dotenv
azure-monitor-opentelemetry==1.8.2
opentelemetry-api==1.39.0
opentelemetry-sdk==1.39.0
